# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.createProviderRequest:
    properties:
      apiKey:
        type: string
      cliArgs:
        items:
          type: string
        type: array
      cliPath:
        type: string
      endpoint:
        type: string
      isDefault:
        type: boolean
      kind:
        description: '"hosted" or "cli"'
        type: string
      model:
        type: string
      name:
        type: string
    required:
    - kind
    - name
    type: object
...
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.openProjectRequest:
    properties:
      path:
        type: string
    required:
    - path
    type: object
  handler.projectResponse:
    properties:
      available:
        type: boolean
      id:
        type: integer
      lastOpenedAt:
        type: string
      name:
        type: string
      path:
        type: string
    type: object
  handler.providerResponse:
    properties:
      cliArgs:
        items:
          type: string
        type: array
      cliPath:
        type: string
      createdAt:
        type: string
      endpoint:
        type: string
      hasApiKey:
        type: boolean
      id:
        type: integer
      isDefault:
        type: boolean
      kind:
        type: string
      model:
        type: string
      name:
        type: string
    type: object
...
  handler.updateProviderRequest:
    properties:
      apiKey:
        type: string
      cliArgs:
        items:
          type: string
        type: array
      cliPath:
        type: string
      endpoint:
        type: string
      model:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
info:
  contact: {}
  description: REST + SSE API for the OpenSpec Studio backend.
  title: OpenSpec Studio API
  version: "1.0"
paths:
  /ai-providers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.providerResponse'
            type: array
      summary: List configured AI providers
      tags:
      - ai-providers
    post:
      consumes:
      - application/json
      parameters:
      - description: Provider config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createProviderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.providerResponse'
        "400":
          description: invalid kind, or CLI executable not found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Configure a new AI provider
      tags:
      - ai-providers
  /ai-providers/{id}:
    delete:
      parameters:
      - description: Provider id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete an AI provider
      tags:
      - ai-providers
    patch:
      consumes:
      - application/json
      parameters:
      - description: Provider id
        in: path
        name: id
        required: true
        type: integer
      - description: Updated config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.providerResponse'
      summary: Update an AI provider
      tags:
      - ai-providers
  /ai-providers/{id}/default:
    post:
      parameters:
      - description: Provider id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Set the default AI provider
      tags:
      - ai-providers
...
  /projects:
    get:
      description: Lists every registered project, most recently opened first. A project
        whose path no longer exists on disk is included with available=false rather
        than omitted.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.projectResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: List registered projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: 'Registers a directory as an active project. Read-only discovery:
        the directory must already contain a valid openspec/ structure - Open never
        scaffolds one. Use POST /projects/new for that.'
      parameters:
      - description: Directory to open
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.openProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.projectResponse'
        "400":
          description: path missing, does not exist, or has no valid openspec/ structure
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Open (register) an existing project
      tags:
      - projects
...
  /projects/new:
    post:
      consumes:
      - application/json
      description: Scaffolds a fresh OpenSpec structure in an empty (or nonexistent-content)
        directory and registers it, distinct from opening an existing project.
      parameters:
      - description: Directory to create the project in
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.openProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.projectResponse'
        "400":
          description: path missing, does not exist, or is not empty
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Create a new project
      tags:
      - projects
...
swagger: "2.0"